feat: listenAndServe returns closer, remote serve()
