+++ router_test.ts
@@ -1,30 +1,13 @@
 // Copyright 2019 Yusuke Sakurai. All rights reserved. MIT license.
-import { createRouter, findLongestAndNearestMatch } from "./router.ts";
-import { runIfMain, test } from "./vendor/https/deno.land/std/testing/mod.ts";
+import { createRouter } from "./router.ts";
+import { runIfMain } from "./vendor/https/deno.land/std/testing/mod.ts";
 import { defer } from "./promises.ts";
 import { assertEquals } from "./vendor/https/deno.land/std/testing/asserts.ts";
 import { it } from "./testing.ts";
 
-([
-  ["/foo", ["/foo", "/bar", "/f"], 0],
-  ["/foo", ["/foo", "/foo/bar"], 0],
-  ["/foo/bar", ["/", "/foo", "/hoo", "/hoo/foo/bar", "/foo/bar"], 4],
-  ["/foo/bar/foo", ["/foo", "/foo/bar", "/bar/foo"], 1],
-  ["/foo", ["/", "/hoo", "/hoo/foo"], 0],
-  ["/deno/land", [/d(.+?)o/, /d(.+?)d/], 1],
-  ["/foo", ["/", "/a/foo"], 0],
-  ["/foo", [/\/foo/, /\/bar\/foo/], 0],
-  ["/foo", [/\/a\/foo/, /\/foo/], 1]
-] as [string, (string | RegExp)[], number][]).forEach(([path, pat, idx]) => {
-  test("findLongestAndNearestMatch:" + path, () => {
-    assertEquals(findLongestAndNearestMatch(path, pat).index, idx);
-  });
-});
-
 it("router", t => {
   let errorHandled = false;
   t.beforeAfterAll(() => {
-    const cancel = defer<void>();
     const router = createRouter();
     router.handle("/index", async req => {
       await req.respond({
@@ -50,14 +33,11 @@ it("router", t => {
     router.handleError((e, req) => {
       errorHandled = true;
     });
-    router.listen(
-      {
-        hostname: "127.0.0.1",
-        port: 8898
-      },
-      { cancel: cancel.promise }
-    );
-    return () => cancel.resolve();
+    const listener = router.listen({
+      hostname: "127.0.0.1",
+      port: 8898
+    });
+    return () => listener.close();
   });
   t.run("should respond string path", async () => {
     {
