+++ tests.ts
@@ -3,6 +3,7 @@
 import "./server_test.ts";
 import "./responder_test.ts";
 import "./router_test.ts";
+import "./router_util_test.ts";
 import "./body_parser_test.ts";
 import "./readers_test.ts";
 import "./serveio_test.ts";
